Create ForgotPasswordForm component to use through the login or register form, or directly through link
